Testing: Refactor support code out of `zope.testing` entrypoint
`tests.py` is the entrypoint file that will be used by `zope.testing` to
discover the test cases on behalf of what's returned from `test_suite`.

It is better to not overload it with other support code that may also be
needed in other contexts.
